Fat16 structure and fopen

+++ include/memory/heap/kheap.h
@@ -5,8 +5,8 @@
 #include <stdint.h>
 
 void kheap_init();
-void* kmalloc(size_t size);
-void* kzalloc(size_t size);
+void* kmalloc(int size);
+void* kzalloc(int size);
 void kfree(void* ptr);
 
 #endif // KHEAP_H
